+++ src/Akeneo/Pim/Structure/Bundle/Controller/InternalApi/AttributeController.php
@@ -16,6 +16,7 @@ use Akeneo\Tool\Component\StorageUtils\Repository\SearchableRepositoryInterface;
 use Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Tool\Component\StorageUtils\Updater\ObjectUpdaterInterface;
 use Akeneo\UserManagement\Bundle\Context\UserContext;
+use Akeneo\UserManagement\Component\Model\UserInterface;
 use Doctrine\Common\Persistence\ObjectRepository;
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -28,6 +29,7 @@ use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Translation\TranslatorInterface;
 use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Webmozart\Assert\Assert;
 
 /**
  * Attribute rest controller
@@ -183,7 +185,9 @@ class AttributeController
         }
 
         $token = $this->tokenStorage->getToken();
-        $options['user_groups_ids'] = $token->getUser()->getGroupsIds();
+        $user = $token->getUser();
+        Assert::implementsInterface($user, UserInterface::class);
+        $options['user_groups_ids'] = $user->getGroupsIds();
 
         $attributes = $this->attributeSearchRepository->findBySearch(
             $request->get('search'),
@@ -467,7 +471,7 @@ class AttributeController
     public function listAxesAction(Request $request)
     {
         $locale = $request->get('locale');
-        $attributeAxes = $this->attributeRepository->getAxesQuery($locale)->getArrayResult();
+        $attributeAxes = $this->attributeRepository->findAvailableAxes($locale);
 
         return new JsonResponse($attributeAxes);
     }